More fixes
Fixed sprites being set to null after changing scenes.
I think I could fix that better, by assigning them to a gameobject, then setting that gameobject to not be unloaded, but idk if that's actually better.

Changed the result screen score rank to be smaller, and at the same position as shinuchi would be, but on the left instead of the right.

Added better utility functions for getting ScoreRank sprite file paths and sprite initialization.